refactor: extract RunwayMiniChart, replace budget-pace with runway chart in Ritmo chip
